Refactored code for rtc implementation
