+++ src/rtc/pages/streaming-page.tsx
@@ -8,7 +8,12 @@ import RtcDynamicVideo from "../components/rtc-dynamic-video";
 import { useEffect, useRef, useState } from "react";
 import { Client, IMessage } from "@stomp/stompjs";
 import RtcToast from "../components/rtc-toast";
-
+import { streamingPageProps } from "../interface/StreamingPage";
+import { makertc } from "../service/makeRtc";
+import { useStompClient } from "@/hooks/useStompClient ";
+import { startCall } from "../service/startCall";
+import { sendSignalingData } from "../service/sendSignalingData";
+import { showToast } from "../service/showToast";
 interface SignalingMessage {
   type: string;
   sdp?: RTCSessionDescriptionInit;
@@ -20,181 +25,219 @@ interface SignalingMessage {
 const StreamingPage: React.FC = () => {
   //* ==========> variables
 
-  //* ==========> refs
-  const stompClientRef = useRef<Client | null>(null);
-  const peerConnectionRef = useRef<RTCPeerConnection | null>(null);
-  const localVideoRef = useRef<HTMLVideoElement>(null!);
-  const remoteVideoRef = useRef<HTMLVideoElement>(null!);
   let popup = false;
-  const localIceCandidateRef = useRef<any>(null);
-  let acceptOrReject: boolean = false;
-  //* ==========> states
-  const [error, setError] = useState<string | null>(null);
-  const [clientIds, setClientIds] = useState<string[]>([]);
-  const [ownId, setownId] = useState<string>("");
-  const targetId = useRef<string | undefined>("");
-  let receivedTargetid: string | undefined = "";
-  let receivedSenderId: string | undefined = "";
   let myId: string | undefined = "";
   let target: string | undefined = "";
+  let acceptOrReject: boolean = false;
+  let receivedSenderId: string | undefined = "";
+  let receivedTargetid: string | undefined = "";
+
+  //* ==========> refs
+  const localIceCandidateRef = useRef<any>(null);
+  const targetId = useRef<string | undefined>("");
+  const stompClientRef = useRef<Client | null>(null);
+  const localVideoRef = useRef<HTMLVideoElement>(null!);
+  const remoteVideoRef = useRef<HTMLVideoElement>(null!);
+  const peerConnectionRef = useRef<RTCPeerConnection | null>(null);
+
+  //* ==========> states
+  const [ownId, setownId] = useState<string>("");
+  const [error, setError] = useState<string | null>(null);
+  const [clientIds, setClientIds] = useState<string[]>([]);
+  const [toasttargetId, setToasttargetId] = useState<string | null>(null);
   const [connectionStatus, setConnectionStatus] = useState<boolean>(false);
   const [isMediaAccessGranted, setIsMediaAccessGranted] =
     useState<boolean>(false);
-  const [toasttargetId, setToasttargetId] = useState<string | null>(null);
-  const showToast = async () => {
-    console.log("receivedTargetid", receivedTargetid, "ownId", ownId);
-    if (popup && myId == receivedTargetid) {
-      toast.custom(
-        (t: Toast) => (
-          <RtcToast
-            onConfirm={async () => {
-              const answer = await peerConnectionRef.current!.createAnswer();
-              await peerConnectionRef.current!.setLocalDescription(answer);
-              if (myId == receivedTargetid) {
-                sendSignalingData({ type: "answer", sdp: answer });
-              }
+  const [ownClientId, setOwnClientId] = useState<string>("");
+  const [client] = useStompClient(
+    peerConnectionRef,
+    setError,
+    streamingPageProps,
+    targetId,
+    ownId,
+    setOwnClientId,
+    setConnectionStatus,
+    stompClientRef,
+    setClientIds
+  );
+  // const showToast = async () => {
+  //   console.log("receivedTargetid", receivedTargetid, "ownId", ownId);
+  //   if (popup && myId == receivedTargetid) {
+  //     toast.custom(
+  //       (t: Toast) => (
+  //         <RtcToast
+  //           onConfirm={async () => {
+  //             const answer = await peerConnectionRef.current!.createAnswer();
+  //             await peerConnectionRef.current!.setLocalDescription(answer);
+  //             if (myId == receivedTargetid) {
+  //               sendSignalingData(
+  //                 { type: "answer", sdp: answer },
+  //                 stompClientRef,
+  //                 streamingPageProps,
+  //                 ownId,
+  //                 targetId
+  //               );
+  //             }
 
-              toast.dismiss(t.id);
-            }}
-            onClose={() => {
-              acceptOrReject = false;
-              toast.dismiss(t.id);
-            }}
-          />
-        ),
-        {
-          duration: Infinity,
-        }
-      );
-    }
-  };
+  //             toast.dismiss(t.id);
+  //           }}
+  //           onClose={() => {
+  //             acceptOrReject = false;
+  //             toast.dismiss(t.id);
+  //           }}
+  //         />
+  //       ),
+  //       {
+  //         duration: Infinity,
+  //       }
+  //     );
+  //   }
+  // };
 
   //* ==========> use-effects
+  showToast(
+    peerConnectionRef,
+    streamingPageProps,
+    ownId,
+    targetId,
+    stompClientRef
+  );
   useEffect(() => {
     // Generate the clientId
-    const clientId = uuidv4();
-    myId = clientId;
-    setownId(clientId);
+    // const clientId = uuidv4();
+    // myId = clientId;
+    // setownId(clientId);
 
     // Include the clientId as a query parameter in the WebSocket URL
-    const websocketUrl = `ws://192.168.1.15:8080/video-websocket?clientId=${clientId}`;
+    // const websocketUrl = `ws://192.168.1.15:8080/video-websocket?clientId=${clientId}`;
 
     // Step 1: Create a STOMP client using native WebSocket with the clientId in the URL
-    const client = new Client({
-      brokerURL: websocketUrl, // Include clientId in the WebSocket URL
-      reconnectDelay: 5000, // Reconnect every 5 seconds if connection is lost
-      heartbeatIncoming: 10000, // Heartbeat interval for incoming messages
-      heartbeatOutgoing: 10000, // Heartbeat interval for outgoing messages
-      debug: (str) => {
-        // console.log("STOMP:", str);
-      },
+    // const client = new Client({
+    //   brokerURL: websocketUrl, // Include clientId in the WebSocket URL
+    //   reconnectDelay: 5000, // Reconnect every 5 seconds if connection is lost
+    //   heartbeatIncoming: 10000, // Heartbeat interval for incoming messages
+    //   heartbeatOutgoing: 10000, // Heartbeat interval for outgoing messages
+    //   debug: (str) => {
+    //     // console.log("STOMP:", str);
+    //   },
 
-      onConnect: () => {
-        setConnectionStatus(true);
-        stompClientRef.current = client;
+    //   onConnect: () => {
+    //     setConnectionStatus(true);
+    //     stompClientRef.current = client;
 
-        client.subscribe("/topic/client-update", (message) => {
-          const updatedSessionMap = JSON.parse(message.body);
-          // console.log("Received updated session map:", updatedSessionMap);
+    //     client.subscribe("/topic/client-update", (message) => {
+    //       const updatedSessionMap = JSON.parse(message.body);
+    //       // console.log("Received updated session map:", updatedSessionMap);
 
-          const ids = Object.keys(updatedSessionMap).filter(
-            (id) => id !== myId
-          );
-          setClientIds(ids);
-        });
+    //       const ids = Object.keys(updatedSessionMap).filter(
+    //         (id) => id !== myId
+    //       );
+    //       setClientIds(ids);
+    //     });
 
-        // Subscribe to receive messages intended for this client
-        client.subscribe(`/user/queue/call`, (message: IMessage) => {
-          try {
-            const data: SignalingMessage = JSON.parse(message.body);
-            console.log("Received signaling data:", data);
-            handleSignalingData(data);
-          } catch (err) {
-            console.error("Error parsing signaling data:", err);
-            setError("Error parsing signaling data");
-          }
-        });
+    //     // Subscribe to receive messages intended for this client
+    //     client.subscribe(`/user/queue/call`, (message: IMessage) => {
+    //       try {
+    //         const data: SignalingMessage = JSON.parse(message.body);
+    //         console.log("Received signaling data:", data);
+    //         handleSignalingData(data);
+    //       } catch (err) {
+    //         console.error("Error parsing signaling data:", err);
+    //         setError("Error parsing signaling data");
+    //       }
+    //     });
 
-        // Send a connect message to the server (Optional if handled during handshake)
-        client.publish({
-          destination: "/app/connect",
-          body: JSON.stringify({ sender: clientId }),
-        });
-      },
-    });
+    //     // Send a connect message to the server (Optional if handled during handshake)
+    //     client.publish({
+    //       destination: "/app/connect",
+    //       body: JSON.stringify({ sender: clientId }),
+    //     });
+    //   },
+    // });
 
     // Step 2: Activate the client to initiate the connection
-    client.activate();
-    const makertc = async () => {
-      const configuration: RTCConfiguration = {
-        iceServers: [
-          { urls: "stun:stun.l.google.com:19302" },
-          // Add TURN servers here if needed
-        ],
-      };
-      const pc = new RTCPeerConnection(configuration);
-      peerConnectionRef.current = pc;
-      pc.onicecandidate = (event: RTCPeerConnectionIceEvent) => {
-        if (event.candidate) {
-          console.log("ICE Candidate found:", event.candidate);
-          localIceCandidateRef.current = event.candidate;
-          console.log("1:", target);
-          if (targetId.current)
-            sendSignalingData({
-              type: "ice-candidate",
-              candidate: event.candidate,
-            });
-        }
-      };
+    if (client) client.activate();
 
-      pc.onicegatheringstatechange = () => {
-        console.log("ICE Gathering State:", pc.iceGatheringState);
-      };
+    // Step 3: Create a WebRTC peer connection
+    // const makertc = async () => {
+    //   const configuration: RTCConfiguration = {
+    //     iceServers: [
+    //       { urls: "stun:stun.l.google.com:19302" },
+    //       // Add TURN servers here if needed
+    //     ],
+    //   };
 
-      // Set remote stream to ref
-      pc.ontrack = (event: RTCTrackEvent) => {
-        console.log("Remote track received:", event);
+    //   const pc = new RTCPeerConnection(configuration);
+    //   peerConnectionRef.current = pc;
 
-        if (!remoteVideoRef.current.srcObject && event.streams.length > 0) {
-          remoteVideoRef.current.srcObject = event.streams[0];
-          console.log(
-            "Remote video stream set",
-            remoteVideoRef.current!.srcObject,
-            event.streams
-          );
-        }
-      };
+    //   pc.onicecandidate = (event: RTCPeerConnectionIceEvent) => {
+    //     if (event.candidate) {
+    //       console.log("ICE Candidate found:", event.candidate);
+    //       localIceCandidateRef.current = event.candidate;
+    //       console.log("1:", target);
+    //       if (targetId.current)
+    //         sendSignalingData({
+    //           type: "ice-candidate",
+    //           candidate: event.candidate,
+    //         });
+    //     }
+    //   };
 
-      // Get user media
-      try {
-        const stream = await navigator.mediaDevices.getUserMedia({
-          video: true,
-          audio: true,
-        });
-        console.log("User media obtained");
-        if (localVideoRef.current) {
-          localVideoRef.current.srcObject = stream;
-          console.log(
-            "Local video stream set",
-            localVideoRef.current.srcObject
-          );
-        }
-        stream.getTracks().forEach((track) => pc.addTrack(track, stream));
-        setIsMediaAccessGranted(true);
-        console.log("Local tracks added to RTCPeerConnection");
-      } catch (mediaError) {
-        setError("Error accessing media devices");
-        setIsMediaAccessGranted(false);
-        console.error("Error accessing media devices.", mediaError);
-      }
-    };
-    // Handle ICE candidates
+    //   pc.onicegatheringstatechange = () => {
+    //     console.log("ICE Gathering State:", pc.iceGatheringState);
+    //   };
 
-    // RTC setup
+    //   pc.ontrack = (event: RTCTrackEvent) => {
+    //     console.log("Remote track received:", event);
 
-    makertc();
+    //     if (!remoteVideoRef.current.srcObject && event.streams.length > 0) {
+    //       remoteVideoRef.current.srcObject = event.streams[0];
+    //       console.log(
+    //         "Remote video stream set",
+    //         remoteVideoRef.current!.srcObject,
+    //         event.streams
+    //       );
+    //     }
+    //   };
 
-    // Cleanup on component unmount
+    //   // Get user media
+    //   try {
+    //     const stream = await navigator.mediaDevices.getUserMedia({
+    //       video: true,
+    //       audio: true,
+    //     });
+    //     console.log("User media obtained");
+    //     if (localVideoRef.current) {
+    //       localVideoRef.current.srcObject = stream;
+    //       console.log(
+    //         "Local video stream set",
+    //         localVideoRef.current.srcObject
+    //       );
+    //     }
+    //     stream.getTracks().forEach((track) => pc.addTrack(track, stream));
+    //     setIsMediaAccessGranted(true);
+    //     console.log("Local tracks added to RTCPeerConnection");
+    //   } catch (mediaError) {
+    //     setError("Error accessing media devices");
+    //     setIsMediaAccessGranted(false);
+    //     console.error("Error accessing media devices.", mediaError);
+    //   }
+    // };
+
+    makertc(
+      peerConnectionRef,
+      localIceCandidateRef,
+      localVideoRef,
+      remoteVideoRef,
+      targetId,
+      target,
+      setIsMediaAccessGranted,
+      stompClientRef,
+      streamingPageProps,
+      ownId
+    );
+
+    // Step 4:  Cleanup on component unmount
     return () => {
       if (client) {
         client.deactivate();
@@ -204,105 +247,105 @@ const StreamingPage: React.FC = () => {
   }, []);
 
   //* ==========> handle functions
-  const handleSignalingData = async (data: SignalingMessage) => {
-    try {
-      const peerConnection = peerConnectionRef.current;
-      if (!peerConnection) {
-        console.error("PeerConnection is not established");
-        setError("PeerConnection is not established");
-        return;
-      }
+  // const handleSignalingData = async (data: SignalingMessage) => {
+  //   try {
+  //     const peerConnection = peerConnectionRef.current;
+  //     if (!peerConnection) {
+  //       console.error("PeerConnection is not established");
+  //       setError("PeerConnection is not established");
+  //       return;
+  //     }
 
-      switch (data.type) {
-        case "offer":
-          popup = true;
-          receivedTargetid = data.target;
-          receivedSenderId = data.sender;
-          console.log("receivedTargetid", receivedTargetid, ownId);
-          targetId.current = data.sender;
-          showToast();
-          await peerConnection.setRemoteDescription(
-            new RTCSessionDescription(data.sdp!)
-          );
+  //     switch (data.type) {
+  //       case "offer":
+  //         popup = true;
+  //         receivedTargetid = data.target;
+  //         receivedSenderId = data.sender;
+  //         console.log("receivedTargetid", receivedTargetid, ownId);
+  //         targetId.current = data.sender;
+  //         showToast();
+  //         await peerConnection.setRemoteDescription(
+  //           new RTCSessionDescription(data.sdp!)
+  //         );
 
-          break;
+  //         break;
 
-        case "answer":
-          await peerConnection.setRemoteDescription(
-            new RTCSessionDescription(data.sdp!)
-          );
-          break;
+  //       case "answer":
+  //         await peerConnection.setRemoteDescription(
+  //           new RTCSessionDescription(data.sdp!)
+  //         );
+  //         break;
 
-        case "ice-candidate":
-          console.log("this will execute");
-          if (data.candidate) {
-            await peerConnection.addIceCandidate(
-              new RTCIceCandidate(data.candidate)
-            );
-          }
-          break;
-      }
-    } catch (error) {
-      console.error("Error handling signaling data:", error);
-    }
-  };
+  //       case "ice-candidate":
+  //         console.log("this will execute");
+  //         if (data.candidate) {
+  //           await peerConnection.addIceCandidate(
+  //             new RTCIceCandidate(data.candidate)
+  //           );
+  //         }
+  //         break;
+  //     }
+  //   } catch (error) {
+  //     console.error("Error handling signaling data:", error);
+  //   }
+  // };
 
-  const sendSignalingData = (data: Partial<SignalingMessage>) => {
-    try {
-      const stompClient = stompClientRef.current;
-      if (stompClient && stompClient.connected) {
-        // Include sender and target in the message
-        const message = {
-          ...data,
-          sender: ownId ? ownId : myId,
-          target:
-            targetId?.current?.length! > 0
-              ? targetId.current
-              : receivedSenderId,
-        };
+  // const sendSignalingData = (data: Partial<SignalingMessage>) => {
+  //   try {
+  //     const stompClient = stompClientRef.current;
+  //     if (stompClient && stompClient.connected) {
+  //       // Include sender and target in the message
+  //       const message = {
+  //         ...data,
+  //         sender: ownId ? ownId : myId,
+  //         target:
+  //           targetId?.current?.length! > 0
+  //             ? targetId.current
+  //             : receivedSenderId,
+  //       };
 
-        stompClient.publish({
-          destination: "/app/call",
-          body: JSON.stringify(message),
-        });
+  //       stompClient.publish({
+  //         destination: "/app/call",
+  //         body: JSON.stringify(message),
+  //       });
 
-        console.log("Sent signaling data:", message);
-      } else {
-        console.error("STOMP client is not connected");
-        setError("STOMP client is not connected");
-      }
-    } catch (error) {
-      console.error("Error sending signaling data:", error);
-      setError("Error sending signaling data");
-    }
-  };
+  //       console.log("Sent signaling data:", message);
+  //     } else {
+  //       console.error("STOMP client is not connected");
+  //       setError("STOMP client is not connected");
+  //     }
+  //   } catch (error) {
+  //     console.error("Error sending signaling data:", error);
+  //     setError("Error sending signaling data");
+  //   }
+  // };
 
-  const startCall = async () => {
-    try {
-      const peerConnection = peerConnectionRef.current;
-      if (!peerConnection) {
-        console.error("PeerConnection is not established");
-        setError("PeerConnection is not established");
-        return;
-      }
+  // const startCall = async () => {
+  //   try {
+  //     const peerConnection = peerConnectionRef.current;
+  //     if (!peerConnection) {
+  //       console.error("PeerConnection is not established");
+  //       setError("PeerConnection is not established");
+  //       return;
+  //     }
 
-      // Create an offer
-      const offer = await peerConnection.createOffer();
-      await peerConnection.setLocalDescription(offer);
-      console.log("Offer created and set as local description", offer);
+  //     // Create an offer
+  //     const offer = await peerConnection.createOffer();
+  //     await peerConnection.setLocalDescription(offer);
+  //     console.log("Offer created and set as local description", offer);
 
-      // Send the offer to the target client
+  //     // Send the offer to the target client
 
-      // if (myId && targetId.current)
-      sendSignalingData({ type: "offer", sdp: offer });
+  //     // if (myId && targetId.current)
+  //     sendSignalingData({ type: "offer", sdp: offer });
 
-      // Show toast only for the target client
-      showToastForClient(targetId.current!);
-    } catch (error) {
-      console.error("Error creating or sending offer:", error);
-      setError("Error creating or sending offer");
-    }
-  };
+  //     // Show toast only for the target client
+  //     showToastForClient(targetId.current!);
+  //   } catch (error) {
+  //     console.error("Error creating or sending offer:", error);
+  //     setError("Error creating or sending offer");
+  //   }
+  // };
 
   const showToastForClient = (targetId: string) => {
     setToasttargetId(targetId);
@@ -312,7 +355,16 @@ const StreamingPage: React.FC = () => {
     <section className="h-screen py-[50px] px-[100px] gap-[30px] flex">
       <RtcSidebar
         clientIds={clientIds}
-        startCall={() => startCall()}
+        startCall={() =>
+          startCall(
+            peerConnectionRef,
+            streamingPageProps,
+            ownId,
+            targetId,
+            stompClientRef,
+            setToasttargetId
+          )
+        }
         ownId={ownId}
         connectionStatus={connectionStatus}
         targetId={targetId}
